coordinator - write secrets to env
This lets us keep all the jobs with `mount: none`, so they can run on
any worker node.

We tie the coordinator to an NFS node, so it has the current credentials
accessible.

The keys are rotated weekly, so we need to do this in 'real time'.

There is a potential race where the key update happens in the middle of
a run, for now let's just call that unlikely.

We could do this before starting each job if needed.